requirements.txt
pandas
openpyxl
pymongo
python-dotenv
flask
flask-cors
gunicorn
numpy
scipy
python-dateutil